ml_service/requirements.txt
flask-restplus==0.13.0
Flask==1.1.2
opencv-contrib-python-headless==4.1.1.26
Pillow==5.4.1
h5py==2.9.0
hdfs==2.5.8
Keras==2.2.4
numpy==1.16.1
pandas==0.24.1
scikit-learn==0.20.2
tensorflow==1.13.0rc2